requirements.txt
Markdown>=3.5.0
python-frontmatter>=1.1.0
Jinja2>=3.1.0
bibtexparser>=1.4.0
beautifulsoup4>=4.12.0
